+++ changelog.md
@@ -1,6 +1,12 @@
 # Changelog
 
 ## [Unreleased]
+### breaking changes
+- rename scope variables for toolbar plugins.
+
+### refactor
+- improve inner code quality on toolbar plugins.
+
 ## [2.3.0]
 ### breaking changes
 - remove mathjs eval by an more accurated eval for js code in toolbar plugins. This will break any plugin created before v2.2.10.